Ability to change the way how time displays through TimeToStringConverter

+++ RecordDataProvider/src/cz/dfi/datamodel/ImageRecordsWrapper.java
@@ -17,7 +17,7 @@ public class ImageRecordsWrapper implements RecordsWrapper {
     }
 
     @Override
-    public double[] getOnBoardTimeValues() {
+    public long[] getOnBoardTimeValues() {
         return null;
     }
