+++ RecordDataProvider/src/cz/dfi/datamodel/graphable/DoubleQuantity.java
@@ -3,8 +3,6 @@
 
 package cz.dfi.datamodel.graphable;
 
-import cz.dfi.datamodel.graphable.AbstractGraphableQuantity;
-
 /**
  *
  * @author Tomas Prochazka
@@ -13,22 +11,22 @@ import cz.dfi.datamodel.graphable.AbstractGraphableQuantity;
 public class DoubleQuantity  extends AbstractGraphableQuantity{
     protected double[] values;
 
-    public DoubleQuantity(double[] values, String name, String unit, double[] onBoardTimeValues) {
+    public DoubleQuantity(double[] values, String name, String unit, long[] onBoardTimeValues) {
         super(name, unit, onBoardTimeValues);
         this.values = values;
     }
 
-    public DoubleQuantity(double[] values, String name, String unit, double[] onBoardTimeValues, boolean isMessageIncoming) {
+    public DoubleQuantity(double[] values, String name, String unit, long[] onBoardTimeValues, boolean isMessageIncoming) {
         super(name, unit, onBoardTimeValues, isMessageIncoming);
         this.values = values;
     }
 
-    public DoubleQuantity(double[] values, String name, String unit, double[] onBoardTimeValues, long[] recorderTimeValues) {
+    public DoubleQuantity(double[] values, String name, String unit, long[] onBoardTimeValues, long[] recorderTimeValues) {
         super(name, unit, onBoardTimeValues, recorderTimeValues);
         this.values = values;
     }
 
-    public DoubleQuantity(double[] values, String name, String unit, double[] onBoardTimeValues, long[] recorderTimeValues, boolean isMessageIncoming) {
+    public DoubleQuantity(double[] values, String name, String unit, long[] onBoardTimeValues, long[] recorderTimeValues, boolean isMessageIncoming) {
         super(name, unit, onBoardTimeValues, recorderTimeValues, isMessageIncoming);
         this.values = values;
     }
