+++ RecordDataProvider/src/cz/dfi/datamodel/graphable/SpeedWrapper.java
@@ -10,7 +10,7 @@ package cz.dfi.datamodel.graphable;
  */
 public abstract class SpeedWrapper extends DoubleQuantity {
 
-    public SpeedWrapper(double[] values, String name, String unit, double[] onBoardTimeValues, long[] recorderTimeValues) {
+    public SpeedWrapper(double[] values, String name, String unit, long[] onBoardTimeValues, long[] recorderTimeValues) {
         super(values, name, unit, onBoardTimeValues, recorderTimeValues);
     }
 }
